+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/implementation/Exceptions.java
@@ -58,4 +58,11 @@ public class Exceptions {
     public static String getInternalServerErrorMessage(String prefix) {
         return prefix + " - " + RMResources.InternalServerError;
     }
+
+    public static boolean isStaledResourceException(int statusCode, int subStatusCode) {
+        return (statusCode == HttpConstants.StatusCodes.BADREQUEST
+                    && subStatusCode == HttpConstants.SubStatusCodes.INCORRECT_CONTAINER_RID_SUB_STATUS)
+            || (statusCode == HttpConstants.StatusCodes.GONE
+                    && subStatusCode == HttpConstants.SubStatusCodes.NAME_CACHE_IS_STALE);
+    }
 }